Add a proper RNG
Many of the better Tetris implementations come with a rather special
RNG.  Initially a bag with seven pieces is creaed and shuffled.  From
now on, every time a piece needs to be spawned, it is popped from the
existing lists until empty, then the cycle can repeat.

The implemented shuffling algorithm is the Knuth shuffle and probably
overkill, but ensures commonly expected behaviur.